requirements.txt
stravalib
python-dotenv
Pillow
matplotlib
geopy
numpy
folium
staticmap
